apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
redis[hiredis]==5.1.1
celery==5.4.0
mediapipe==0.10.14
openai==1.51.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
httpx==0.27.2
python-multipart==0.0.12
pydantic==2.9.2
